Package updates

+++ tests/Equinox.EventStore.Integration/EventStoreTokenTests.fs
@@ -37,7 +37,7 @@ let ``ofUncompactedVersion - batchCapacityLimit`` streamVersion batchSize expect
     ; InlineData(   0,  2, 1, 3, 0)
     ; InlineData(   1,  2, 1, 3, 0)
     ; InlineData(   2,  2, 1, 3, 1)>]
-let ``ofPreviousTokenAndEventsLength - batchCapacityLimit`` (previousCompactionEventNumber : System.Nullable<int64>) streamVersion eventsLength batchSize expectedCapacity =
+let ``ofPreviousTokenAndEventsLength - batchCapacityLimit`` (previousCompactionEventNumber : System.Nullable<int>) streamVersion eventsLength batchSize expectedCapacity =
     let previousToken =
         if not previousCompactionEventNumber.HasValue then Token.ofCompactionEventNumber None 0 -84 -42L
         else Token.ofCompactionEventNumber (Some previousCompactionEventNumber.Value) 0 -84 -42L
